AuthService was changed. Logging in Slack
